Add GitHub authentication
I feel that GitHub is one of the easiest and friendly OAuth system.
Integration was easy, but passportjs documentation is a mess :(
Spent a lot of time figuring out serialization/deserialization
of the user.

Why separate layout for login page? I feel that it would be cruel
to send the user straight away to the github authorization page.
The intermediate page allow user to think and decide.