Add post save signal for the User creation
Took me a bit of time to figure it out, especially the
await part. It is kinda obvious, but I just forgot thatI am in the
asyncronous land.

Anyway this logic is here because I don't want user to start from
the empty space, that way he can start with renaming habits and
editing them as helike instead of figuring out how to create one.