Add first API endpoint for habits
First step is to be able to get a list of habits from the API.
It is really nice that since it is a one server cookies gets sended
with the request and I have a `req.user` inside the route!
That way I am avoid sending the user/session/JWT token explicitly
to the API and doing user fetching.