Fix sourcemap generation
So, there is a bug in the parcel-bundler. I was generating the
sourcemap with a wrong path, but in reality it should've been
relative to the client code. I.e. my bundle in `/public/client.js`
but inside that file the last string is a link to the source maps
and it was `/client.js.map`, and of course Firefox was unable to
find that file.

https://github.com/parcel-bundler/parcel/issues/1028