Add styleguide, JSX & HabitListView
As client started to kick off, I wanted to have a styleguide
for the project in order to better understand the UI
elements that I will be using and trying to minimize the CSS.
With minimizing the CSS I failed miserably, not happy with
the things as they are right now. With styleguide I am
using static HTML, because I installed the storybook,
it started complaining about some config files, won't work
with the Mithril, yada yada. In the end uninstalled it
and was horrified that 490 dependencies was neede to
run Storybook.

JSX. I started with Mithril template syntax, but it was
a bit funky. I am pretty sure that adding JSX was not
necesarry, and that bundle building time increased by
3x. But for now it feels more natural and easier to read.