Change Mithril to React
Previously I started writing the application in Mithril,
but when I came to forms, I stubled in strange logic
and spent some time fighting it to make it work. In the
end as I use React in my work, I thought that for this
project I can try to imlement it first in React and see
where this will get us. One thing that I am not happy
at all is that now compile time is almost 10s :(