Testing: Add Jest testing framework
It was a painful experience. First I tried to use mocha,
it is a fast framework but I could not make it work with testing
requests to the server for the pages :( then I tried out Jest,
did not worked out too, then I rewrote useEffect components to
componentDidMount stuff and it was almost running. In the end
I added a sleep function and I did not liked it at all. It seems
to me that in JS testing framworks are still immature, and
easy task to test the request  to the server is making a fool
of me :(