Swap React with Preact
Frankly, I was hoping for more performant code :( The build process
even with hyperscript is aroud 8s for me, which is unaceptable.
But the biggest problem is that it seems that nobody in Preact
community is using Preact without JSX, i.e. Preact is just
swallowing all the template errors, the console is clear :(
There is no way to found out where it fails to parse the markup.
I had very high expectations and in return got very dissapointing
results.